WQS-115: Move drupal test helper to separate directory
